Revert "modify the python files in path_follower package to debug the matfile…"

+++ src/simu/include/f_6s.h
@@ -3,5 +3,5 @@
 
 #include<iostream>
 using namespace std;
-float *f_6s(float* ,float* ,float* ,float* ,float* ,float ,float );
+double *f_6s(double* ,double* ,double* ,double* ,double* ,double ,double );
 #endif
